amdgpuadapterhandle
